+++ packages/local_auth/local_auth_platform_interface/lib/default_method_channel_platform.dart
@@ -17,13 +17,12 @@ const MethodChannel _channel = MethodChannel('plugins.flutter.io/local_auth');
 class DefaultLocalAuthPlatform extends LocalAuthPlatform {
   @override
   Future<bool> authenticate({
-    required String localizedReason,
+    String? localizedReason,
     required Iterable<AuthMessages> authMessages,
     AuthenticationOptions options = const AuthenticationOptions(),
   }) async {
-    assert(localizedReason.isNotEmpty);
     final args = <String, Object>{
-      'localizedReason': localizedReason,
+      if (localizedReason != null) 'localizedReason': localizedReason,
       'useErrorDialogs': options.useErrorDialogs,
       'stickyAuth': options.stickyAuth,
       'sensitiveTransaction': options.sensitiveTransaction,